// __generated__/globalTypes.ts
export interface UpdateAreaInput {
  id: string;
  name?: string | null;
  survey_id?: string | null;
  friendly_token?: string | null;
  clientMutationId?: string | null;
}